Move PCR to utils module in favor of PCR comparison
