+++ src/scib_metrics/__init__.py
@@ -5,7 +5,7 @@ from . import utils
 from ._ari_nmi import nmi_ari_cluster_labels_kmeans, nmi_ari_cluster_labels_leiden
 from ._isolated_labels import isolated_labels
 from ._lisi import clisi_knn, ilisi_knn, lisi_knn
-from ._pcr import pcr_comparison, principal_component_regression
+from ._pcr_comparison import pcr_comparison
 from ._settings import settings
 from ._silhouette import silhouette_batch, silhouette_label
 
@@ -13,7 +13,6 @@ __all__ = [
     "utils",
     "isolated_labels",
     "pcr_comparison",
-    "principal_component_regression",
     "silhouette_label",
     "silhouette_batch",
     "ilisi_knn",
