+++ src/scib_metrics/utils/__init__.py
@@ -2,6 +2,7 @@ from ._dist import cdist, pdist_squareform
 from ._kmeans import KMeansJax
 from ._lisi import compute_simpson_index
 from ._pca import pca
+from ._pcr import principal_component_regression
 from ._silhouette import silhouette_samples
 from ._utils import get_ndarray, one_hot
 
@@ -12,6 +13,7 @@ __all__ = [
     "get_ndarray",
     "KMeansJax",
     "pca",
+    "principal_component_regression",
     "one_hot",
     "compute_simpson_index",
 ]
