+++ tests/utils/test_pcr.py
@@ -12,7 +12,6 @@ from tests.utils.sampling import categorical_sample, normal_sample, poisson_samp
 PCR_PARAMS = list(product([10, 100, 1000], [10, 100, 1000], [False]))
 # TODO(martinkim0): Currently not testing categorical covariates because of
 # TODO(martinkim0): reproducibility issues with original scib. See comment in PR #16.
-PCR_COMPARISON_PARAMS = list(product([100], [100], [False, True]))
 
 
 @pytest.mark.parametrize("n_obs, n_vars, categorical", PCR_PARAMS)
@@ -39,13 +38,3 @@ def test_pcr(n_obs, n_vars, categorical):
     _test_pcr(n_obs, n_vars, n_components=max_components - 1, categorical=categorical)
     _test_pcr(n_obs, n_vars, n_components=int(max_components / 2), categorical=categorical)
     _test_pcr(n_obs, n_vars, n_components=1, categorical=categorical)
-
-
-@pytest.mark.parametrize("n_obs, n_vars, categorical", PCR_COMPARISON_PARAMS)
-def test_pcr_comparison(n_obs, n_vars, categorical):
-    X_pre = poisson_sample(n_obs, n_vars, seed=0)
-    X_post = poisson_sample(n_obs, n_vars, seed=1)
-    covariate = categorical_sample(n_obs, int(n_obs / 5)) if categorical else normal_sample(n_obs, seed=0)
-
-    score = scib_metrics.pcr_comparison(X_pre, X_post, covariate, scale=True)
-    assert score >= 0 and score <= 1
